+++ test/compute-engine/incomplete-expressions.test.ts
@@ -12,45 +12,29 @@ describe('basic', () => {
     expect(parse('\\frac{')).toMatchInlineSnapshot(`
       [
         "Divide",
-        {
-          fn: ["Error", "expected-closing-delimiter", ["LatexString", "{"]];
-            sourceOffsets: [5, 6]
-        },
-        {fn: ["Error", "'missing'"]; sourceOffsets: [6, 6]}
+        ["Error", "expected-closing-delimiter", ["LatexString", "{"]],
+        ["Error", "'missing'"]
       ]
     `);
     expect(parse('\\frac{{')).toMatchInlineSnapshot(`
       [
         "Divide",
-        {
-          fn: ["Error", "expected-closing-delimiter", ["LatexString", "{"]];
-            sourceOffsets: [6, 7]
-        },
-        {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]}
+        ["Error", "expected-closing-delimiter", ["LatexString", "{"]],
+        ["Error", "'missing'"]
       ]
     `);
     expect(parse('\\frac{}}')).toMatchInlineSnapshot(`
       [
         "Tuple",
-        [
-          "Divide",
-          {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]},
-          {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]}
-        ],
-        {
-          fn: ["Error", "unexpected-closing-delimiter", ["LatexString", "}"]];
-            sourceOffsets: [7, 8]
-        }
+        ["Divide", ["Error", "'missing'"], ["Error", "'missing'"]],
+        ["Error", "unexpected-closing-delimiter", ["LatexString", "}"]]
       ]
     `);
     expect(parse('\\frac{1}}')).toMatchInlineSnapshot(`
       [
         "Tuple",
-        ["Divide", 1, {fn: ["Error", "'missing'"]; sourceOffsets: [8, 8]}],
-        {
-          fn: ["Error", "unexpected-closing-delimiter", ["LatexString", "}"]];
-            sourceOffsets: [8, 9]
-        }
+        ["Divide", 1, ["Error", "'missing'"]],
+        ["Error", "unexpected-closing-delimiter", ["LatexString", "}"]]
       ]
     `);
     expect(parse('\\frac{1}{2')).toMatchInlineSnapshot(`
@@ -60,81 +44,54 @@ describe('basic', () => {
         [
           "Tuple",
           2,
-          {
-            fn: ["Error", "expected-closing-delimiter", ["LatexString", "{2"]];
-                sourceOffsets: [8, 10]
-          }
+          ["Error", "expected-closing-delimiter", ["LatexString", "{2"]]
         ]
       ]
     `);
     expect(parse('\\sqrt{}')).toMatchInlineSnapshot(
-      `["Sqrt", {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]}]`
+      `["Sqrt", ["Error", "'missing'"]]`
     );
     expect(parse('\\sqrt{}{}')).toMatchInlineSnapshot(
-      `["Sqrt", {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]}]`
+      `["Sqrt", ["Error", "'missing'"]]`
     );
     expect(parse('\\sqrt')).toMatchInlineSnapshot(
-      `["Sqrt", {fn: ["Error", "'missing'"]; sourceOffsets: [5, 5]}]`
+      `["Sqrt", ["Error", "'missing'"]]`
     );
   });
 
   test('Semantic Errors', () => {
-    expect(parse('1+')).toMatchInlineSnapshot(`
-      [
-        "Sequence",
-        1,
-        {
-          fn: ["Error", "unexpected-operator", ["LatexString", "+"]];
-            sourceOffsets: [1, 2]
-        }
-      ]
-    `);
+    expect(parse('1+')).toMatchInlineSnapshot(
+      `["Sequence", 1, ["Error", "unexpected-operator", ["LatexString", "+"]]]`
+    );
     expect(parse('1\\times')).toMatchInlineSnapshot(
       `["Multiply", 1, ["Error", "'missing'"]]`
     );
-    expect(parse('\\times')).toMatchInlineSnapshot(`
-      {
-        fn: ["Error", "unexpected-command", ["LatexString", "\\times"]];
-        sourceOffsets: [0, 6]
-      }
-    `);
+    expect(parse('\\times')).toMatchInlineSnapshot(
+      `["Error", "unexpected-command", ["LatexString", "\\times"]]`
+    );
     expect(parse('\\times3')).toMatchInlineSnapshot(`
       [
         "Tuple",
-        {
-          fn: ["Error", "unexpected-command", ["LatexString", "\\times"]];
-            sourceOffsets: [0, 6]
-        },
+        ["Error", "unexpected-command", ["LatexString", "\\times"]],
         3
       ]
     `);
     expect(parse('2*')).toMatchInlineSnapshot(
       `["Multiply", 2, ["Error", "'missing'"]]`
     );
-    expect(parse('\\frac{}{}')).toMatchInlineSnapshot(`
-      [
-        "Divide",
-        {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]},
-        {fn: ["Error", "'missing'"]; sourceOffsets: [9, 9]}
-      ]
-    `);
+    expect(parse('\\frac{}{}')).toMatchInlineSnapshot(
+      `["Divide", ["Error", "'missing'"], ["Error", "'missing'"]]`
+    );
     expect(parse('\\frac{1}{}')).toMatchInlineSnapshot(
-      `["Divide", 1, {fn: ["Error", "'missing'"]; sourceOffsets: [10, 10]}]`
+      `["Divide", 1, ["Error", "'missing'"]]`
     );
     expect(parse('\\frac{}{2}')).toMatchInlineSnapshot(
-      `["Divide", {fn: ["Error", "'missing'"]; sourceOffsets: [7, 7]}, 2]`
+      `["Divide", ["Error", "'missing'"], 2]`
     );
 
-    expect(parse('=x')).toMatchInlineSnapshot(`
-      [
-        "Equal",
-        {
-          fn: ["Error", "'missing'", ["LatexString", "="]];
-            sourceOffsets: [1, 1]
-        },
-        "x"
-      ]
-    `);
+    expect(parse('=x')).toMatchInlineSnapshot(
+      `["Equal", ["Error", "'missing'", ["LatexString", "="]], "x"]`
+    );
   });
 
   test('Valid Incomplete', () => {